question-generation-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==3.1.1
scikit-learn==1.5.2
numpy==2.1.1